Fixed the temperature conversion errors.

+++ src/com/umuzi/TaskSeven.java
@@ -3,13 +3,13 @@ package com.umuzi;
 public class TaskSeven {
 
     public double celsiusToFahrenheit(double temperature){
-        double temperatureInFahrenheit = (temperature * (9/5)) + 32;
+        double temperatureInFahrenheit = (temperature * (double) 9 / 5) + 32;
 
         return temperatureInFahrenheit;
     }
 
     public double fahrenheitToCelsius(double temperature) {
-        double temperatureInFahrenheit = (temperature - 32) * (5 / 9);
+        double temperatureInFahrenheit = (temperature - 32) * (double)5 / 9;
 
         return temperatureInFahrenheit;
     }
